+++ dspy/adapters/__init__.py
@@ -1,8 +1,7 @@
 from dspy.adapters.base import Adapter
 from dspy.adapters.chat_adapter import ChatAdapter
 from dspy.adapters.json_adapter import JSONAdapter
-from dspy.adapters.image_utils import Image
-from dspy.adapters.types import History
+from dspy.adapters.types import Image, History
 
 __all__ = [
     "Adapter",
