+++ dspy/adapters/types/__init__.py
@@ -1,3 +1,4 @@
 from dspy.adapters.types.history import History
+from dspy.adapters.types.image import Image
 
-__all__ = ["History"]
+__all__ = ["History", "Image"]
