+++ dspy/adapters/types/image.py
