+++ dspy/signatures/signature.py
@@ -27,7 +27,7 @@ from typing import Any, Dict, Tuple, Type, Union  # noqa: UP035
 from pydantic import BaseModel, Field, create_model
 from pydantic.fields import FieldInfo
 
-from dspy.adapters.image_utils import Image  # noqa: F401
+from dspy.adapters.types.image import Image  # noqa: F401
 from dspy.signatures.field import InputField, OutputField
